move ris_loader utils
