+++ colrev/ops/built_in/search_sources/trid.py
@@ -12,7 +12,7 @@ from dataclasses_jsonschema import JsonSchemaMixin
 
 import colrev.env.package_manager
 import colrev.exceptions as colrev_exceptions
-import colrev.ops.built_in.search_sources.ris_utils
+import colrev.ops.load_utils_ris
 import colrev.ops.search
 import colrev.record
 
@@ -118,13 +118,11 @@ class TransportResearchInternationalDocumentation(JsonSchemaMixin):
         """Load the records from the SearchSource file"""
 
         if self.search_source.filename.suffix == ".ris":
-            ris_entries = colrev.ops.built_in.search_sources.ris_utils.load_ris_entries(
+            ris_entries = colrev.ops.load_utils_ris.load_ris_entries(
                 filename=self.search_source.filename
             )
             self.__ris_fixes(entries=ris_entries)
-            records = colrev.ops.built_in.search_sources.ris_utils.convert_to_records(
-                ris_entries
-            )
+            records = colrev.ops.load_utils_ris.convert_to_records(ris_entries)
             return records
 
         raise NotImplementedError
