--- docs/source/foundations/_autosummary/colrev.ops.built_in.load_conversion.zotero_loader.ZoteroTranslationLoader.rst
@@ -1,39 +1,0 @@
-colrev.ops.built\_in.load\_conversion.zotero\_loader.ZoteroTranslationLoader
-============================================================================
-
-.. currentmodule:: colrev.ops.built_in.load_conversion.zotero_loader
-
-.. autoclass:: ZoteroTranslationLoader
-   :members:
-   :show-inheritance:
-   :inherited-members:
-   :special-members: __call__, __add__, __mul__
-
-
-
-   .. rubric:: Methods
-
-   .. autosummary::
-      :nosignatures:
-
-      ~ZoteroTranslationLoader.all_json_schemas
-      ~ZoteroTranslationLoader.field_mapping
-      ~ZoteroTranslationLoader.from_dict
-      ~ZoteroTranslationLoader.from_json
-      ~ZoteroTranslationLoader.from_object
-      ~ZoteroTranslationLoader.json_schema
-      ~ZoteroTranslationLoader.load
-      ~ZoteroTranslationLoader.register_field_encoders
-      ~ZoteroTranslationLoader.to_dict
-      ~ZoteroTranslationLoader.to_json
-
-
-
-
-
-   .. rubric:: Attributes
-
-   .. autosummary::
-
-      ~ZoteroTranslationLoader.ci_supported
-      ~ZoteroTranslationLoader.supported_extensions
