--- docs/source/foundations/_autosummary/colrev.ops.built_in.load_conversion.zotero_loader.rst
@@ -1,23 +1,0 @@
-colrev.ops.built\_in.load\_conversion.zotero\_loader
-====================================================
-
-.. automodule:: colrev.ops.built_in.load_conversion.zotero_loader
-
-
-
-
-
-
-
-
-
-
-
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree:
-      :template: custom-class-template.rst
-      :nosignatures:
-
-      ZoteroTranslationLoader
