--- docs/source/foundations/_autosummary/colrev.ops.built_in.search_sources.ris_utils.apply_ris_fixes.rst
@@ -1,6 +1,0 @@
-colrev.ops.built\_in.search\_sources.ris\_utils.apply\_ris\_fixes
-=================================================================
-
-.. currentmodule:: colrev.ops.built_in.search_sources.ris_utils
-
-.. autofunction:: apply_ris_fixes
