--- docs/source/foundations/_autosummary/colrev.ops.built_in.search_sources.ris_utils.convert_to_records.rst
@@ -1,6 +1,0 @@
-colrev.ops.built\_in.search\_sources.ris\_utils.convert\_to\_records
-====================================================================
-
-.. currentmodule:: colrev.ops.built_in.search_sources.ris_utils
-
-.. autofunction:: convert_to_records
