--- docs/source/foundations/_autosummary/colrev.ops.built_in.search_sources.ris_utils.load_ris_entries.rst
@@ -1,6 +1,0 @@
-colrev.ops.built\_in.search\_sources.ris\_utils.load\_ris\_entries
-==================================================================
-
-.. currentmodule:: colrev.ops.built_in.search_sources.ris_utils
-
-.. autofunction:: load_ris_entries
