+++ docs/source/foundations/_autosummary/colrev.ops.built_in.search_sources.rst
@@ -48,7 +48,6 @@ colrev.ops.built\_in.search\_sources
    colrev.ops.built_in.search_sources.pdfs_dir
    colrev.ops.built_in.search_sources.psycinfo
    colrev.ops.built_in.search_sources.pubmed
-   colrev.ops.built_in.search_sources.ris_utils
    colrev.ops.built_in.search_sources.scopus
    colrev.ops.built_in.search_sources.springer_link
    colrev.ops.built_in.search_sources.synergy_datasets
