+++ docs/source/foundations/_autosummary/colrev.ops.load_utils_ris.apply_ris_fixes.rst
@@ -1,0 +1,6 @@
+colrev.ops.load\_utils\_ris.apply\_ris\_fixes
+=============================================
+
+.. currentmodule:: colrev.ops.load_utils_ris
+
+.. autofunction:: apply_ris_fixes
