+++ docs/source/foundations/_autosummary/colrev.ops.load_utils_ris.convert_to_records.rst
@@ -1,0 +1,6 @@
+colrev.ops.load\_utils\_ris.convert\_to\_records
+================================================
+
+.. currentmodule:: colrev.ops.load_utils_ris
+
+.. autofunction:: convert_to_records
