+++ docs/source/foundations/_autosummary/colrev.ops.load_utils_ris.load_ris_entries.rst
@@ -1,0 +1,6 @@
+colrev.ops.load\_utils\_ris.load\_ris\_entries
+==============================================
+
+.. currentmodule:: colrev.ops.load_utils_ris
+
+.. autofunction:: load_ris_entries
